update on backend

+++ backend/dilivary/dilivary/vehicule/models.py
@@ -32,7 +32,7 @@ class Vehicle(models.Model):
     matricule = models.IntegerField(primary_key=True)
     color = models.CharField(max_length=50)
     type = models.CharField(max_length=20, choices=VEHICLE_TYPES)
-    max_weight = models.DecimalField(max_digits=10, decimal_places=2)
+    max_weight = models.IntegerField()
     max_size = models.CharField(max_length=20, choices=MAX_SIZE_CHOICES)
     id_driving_license = models.IntegerField(unique=True)
     model = models.CharField(max_length=50)
